He hecho hasta el ejercicio 12 de C#

+++ Ejercicios de C#/Ejercicio3/Ejercicio3/Program.cs
@@ -18,7 +18,7 @@ namespace Ejercicio3
 				i++;
 			} while (i <= number) ;
 
-			double avg = sum / number;
+			double avg = (double)sum / (double)number;
 
 			Console.WriteLine ("The sum is: " + sum);
 			Console.WriteLine ("The average is: " + avg);
